Fixed several bugs flushed out from the black box tests.
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/array-rework@3312 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libCore/SVD.cpp
@@ -999,8 +999,8 @@ ArrayVector SVDFunction(int nargout, const ArrayVector& arg) {
       return SVDFunction(A.real<float>(),A.imag<float>(),computevectors,compactform);
   case Double:
     if (A.allReal())
-      return SVDFunction(A.real<float>(),computevectors,compactform);
+      return SVDFunction(A.real<double>(),computevectors,compactform);
     else
-      return SVDFunction(A.real<float>(),A.imag<float>(),computevectors,compactform);
+      return SVDFunction(A.real<double>(),A.imag<double>(),computevectors,compactform);
   }
 }
